FIX #7806 reset preferences

+++ src/app/basket/models/GroupBasketModelAbstract.php
@@ -4,11 +4,11 @@
  * Copyright Maarch since 2008 under licence GPLv3.
  * See LICENCE.txt file at the root folder for more details.
  * This file is part of Maarch software.
+ *
  */
 
 /**
  * @brief   GroupBasket Model Abstract
- *
  * @author  dev@maarch.org
  */
 
@@ -25,11 +25,11 @@ abstract class GroupBasketModelAbstract
         ValidatorModel::arrayType($aArgs, ['select', 'where', 'data', 'orderBy']);
 
         $aGroupsBaskets = DatabaseModel::select([
-            'select' => empty($aArgs['select']) ? ['*'] : $aArgs['select'],
-            'table' => ['groupbasket'],
-            'where' => $aArgs['where'],
-            'data' => $aArgs['data'],
-            'order_by' => $aArgs['orderBy'],
+            'select'    => empty($aArgs['select']) ? ['*'] : $aArgs['select'],
+            'table'     => ['groupbasket'],
+            'where'     => $aArgs['where'],
+            'data'      => $aArgs['data'],
+            'order_by'  => $aArgs['orderBy']
         ]);
 
         return $aGroupsBaskets;
@@ -41,12 +41,12 @@ abstract class GroupBasketModelAbstract
         ValidatorModel::stringType($aArgs, ['basketId', 'groupId', 'resultPage']);
 
         DatabaseModel::insert([
-            'table' => 'groupbasket',
+            'table'         => 'groupbasket',
             'columnsValues' => [
-                'basket_id' => $aArgs['basketId'],
-                'group_id' => $aArgs['groupId'],
-                'result_page' => $aArgs['resultPage'],
-            ],
+                'basket_id'         => $aArgs['basketId'],
+                'group_id'          => $aArgs['groupId'],
+                'result_page'       => $aArgs['resultPage']
+            ]
         ]);
 
         return true;
@@ -56,65 +56,38 @@ abstract class GroupBasketModelAbstract
     {
         ValidatorModel::notEmpty($aArgs, ['basketId', 'groupId']);
         ValidatorModel::stringType($aArgs, ['basketId', 'groupId']);
+        ValidatorModel::boolType($aArgs, ['preferences']);
 
         $group = GroupModel::getByGroupId(['select' => ['id'], 'groupId' => $aArgs['groupId']]);
 
         DatabaseModel::delete([
             'table' => 'groupbasket',
             'where' => ['basket_id = ?', 'group_id = ?'],
-            'data' => [$aArgs['basketId'], $aArgs['groupId']],
+            'data'  => [$aArgs['basketId'], $aArgs['groupId']]
         ]);
         DatabaseModel::delete([
             'table' => 'actions_groupbaskets',
             'where' => ['basket_id = ?', 'group_id = ?'],
-            'data' => [$aArgs['basketId'], $aArgs['groupId']],
+            'data'  => [$aArgs['basketId'], $aArgs['groupId']]
         ]);
         DatabaseModel::delete([
             'table' => 'groupbasket_redirect',
             'where' => ['basket_id = ?', 'group_id = ?'],
-            'data' => [$aArgs['basketId'], $aArgs['groupId']],
+            'data'  => [$aArgs['basketId'], $aArgs['groupId']]
         ]);
         DatabaseModel::delete([
             'table' => 'groupbasket_status',
             'where' => ['basket_id = ?', 'group_id = ?'],
-            'data' => [$aArgs['basketId'], $aArgs['groupId']],
-        ]);
-        DatabaseModel::delete([
-            'table' => 'users_baskets_preferences',
-            'where' => ['basket_id = ?', 'group_serial_id = ?'],
-            'data' => [$aArgs['basketId'], $group['id']],
+            'data'  => [$aArgs['basketId'], $aArgs['groupId']]
         ]);
 
-        return true;
-    }
-
-    public static function deleteGroupBasketWithoutUserPref(array $aArgs)
-    {
-        ValidatorModel::notEmpty($aArgs, ['basketId', 'groupId']);
-        ValidatorModel::stringType($aArgs, ['basketId', 'groupId']);
-
-        $group = GroupModel::getByGroupId(['select' => ['id'], 'groupId' => $aArgs['groupId']]);
-
-        DatabaseModel::delete([
-            'table' => 'groupbasket',
-            'where' => ['basket_id = ?', 'group_id = ?'],
-            'data' => [$aArgs['basketId'], $aArgs['groupId']],
-        ]);
-        DatabaseModel::delete([
-            'table' => 'actions_groupbaskets',
-            'where' => ['basket_id = ?', 'group_id = ?'],
-            'data' => [$aArgs['basketId'], $aArgs['groupId']],
-        ]);
-        DatabaseModel::delete([
-            'table' => 'groupbasket_redirect',
-            'where' => ['basket_id = ?', 'group_id = ?'],
-            'data' => [$aArgs['basketId'], $aArgs['groupId']],
-        ]);
-        DatabaseModel::delete([
-            'table' => 'groupbasket_status',
-            'where' => ['basket_id = ?', 'group_id = ?'],
-            'data' => [$aArgs['basketId'], $aArgs['groupId']],
-        ]);
+        if (!empty($aArgs['preferences'])) {
+            DatabaseModel::delete([
+                'table' => 'users_baskets_preferences',
+                'where' => ['basket_id = ?', 'group_serial_id = ?'],
+                'data'  => [$aArgs['basketId'], $group['id']]
+            ]);
+        }
 
         return true;
     }
